Refactor dialog input
